+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add stats to `nextify`. ([@palkan][])
+
 - Add text rewriters and `--import-rewriter` option to `nextify`. ([@palkan][])
 
 - Use `.rbnextrc` arguments for runtime transpiling and auto-transpiling gems. ([@palkan][])